Phonetic substitutions for the count
This hack inserts one or two small regular expressions that rewrite the
lastUpdate string with more phonetically natural representations. The
count (balls and strikes) of an at bat appearance, written e.g. "0-2"
is generally pronounced "oh and two" or simply "oh two." TTS engines
like to pronounce the hyphen as "to" because of the character's
association with numerical ranges in programming, but this jarringly
makes the phrase sound like a score when spoken.

Dashes in the count will be replaced with ' and ' (including spaces);
zeroes in the count will be replaced with the letter 'O'. The phonetic
representation will be printed to the console at the debug level.

If there's a better way to do this, I don't know what it is because I am
rubbish at JavaScript, but it does at least work and sound the way my
pedantic brain wants it to.